Reduce uint64 to uint32 for easy conversion with float64.

+++ pkg/api/client/check.go
@@ -13,7 +13,7 @@ func NewCheckClient(c ScoretrakClient) check.Serv {
 	return &CheckClient{c}
 }
 
-func (s CheckClient) GetAllByRoundID(rID uint64) ([]*check.Check, error) {
+func (s CheckClient) GetAllByRoundID(rID uint32) ([]*check.Check, error) {
 	var chk []*check.Check
 	err := s.s.GenericGet(&chk, fmt.Sprintf("/check/%d", rID))
 	if err != nil {
@@ -22,7 +22,7 @@ func (s CheckClient) GetAllByRoundID(rID uint64) ([]*check.Check, error) {
 	return chk, nil
 }
 
-func (s CheckClient) GetByRoundServiceID(rID uint64, sID uint64) (*check.Check, error) {
+func (s CheckClient) GetByRoundServiceID(rID uint32, sID uint32) (*check.Check, error) {
 	var chk *check.Check
 	err := s.s.GenericGet(&chk, fmt.Sprintf("/check/%d/%d", rID, sID))
 	if err != nil {
