+++ pkg/service/serv.go
@@ -1,9 +1,9 @@
 package service
 
 type Serv interface {
-	Delete(id uint64) error
+	Delete(id uint32) error
 	GetAll() ([]*Service, error)
-	GetByID(id uint64) (*Service, error)
+	GetByID(id uint32) (*Service, error)
 	Store(u *Service) error
 	Update(u *Service) error
 }
@@ -18,11 +18,11 @@ func NewServiceServ(repo Repo) Serv {
 	}
 }
 
-func (svc *serviceServ) Delete(id uint64) error { return svc.repo.Delete(id) }
+func (svc *serviceServ) Delete(id uint32) error { return svc.repo.Delete(id) }
 
 func (svc *serviceServ) GetAll() ([]*Service, error) { return svc.repo.GetAll() }
 
-func (svc *serviceServ) GetByID(id uint64) (*Service, error) { return svc.repo.GetByID(id) }
+func (svc *serviceServ) GetByID(id uint32) (*Service, error) { return svc.repo.GetByID(id) }
 
 func (svc *serviceServ) Store(u *Service) error { return svc.repo.Store(u) }
